refactor(bib): move `.bibliography` to stdlib `Bibliography` module
